fix(capx): 清理 8 个 pre-existing clippy warning
- canvas_heal.rs make_frame_dark_editor doc：列表项后加空行隔开普通段落
  （修 2 个 doc_lazy_continuation + 1 个 lib.rs:1 attribute 提示）
- canvas_heal.rs 3 处闭包 noise 计算：去掉冗余 x as u32 / y as u32
  （x/y 来自 for x in 0..TW，TW 是 u32，已是 u32——修 6 个 unnecessary_cast）

cargo clippy -p octopus-capx --all-targets → 0 warning（原 8 → 0）
cargo test -p octopus-capx → 55 passed（行为不变）

+++ crates/capx/src/stitch/canvas_heal.rs
@@ -291,11 +291,12 @@ mod tests {
     }
 
     /// 合成「暗色代码编辑器」帧：近黑背景 + 稀疏亮文字行（等宽字体感）。
+    ///
     /// - 背景 luma≈12（近黑）
     /// - 行周期 24px：16px 文字行 + 8px 纯黑行间
     /// - 文字行内字符周期 11px（6px 亮 luma=220 + 5px 黑），模拟代码字符
-    /// `scroll_offset` 模拟向下滚动。
-    /// 复刻真实暗色编辑器：高灰度对比但 Sobel 特征稀疏（大片纯黑行间）。
+    ///
+    /// `scroll_offset` 模拟向下滚动。复刻真实暗色编辑器：高灰度对比但 Sobel 特征稀疏（大片纯黑行间）。
     fn make_frame_dark_editor(width: u32, height: u32, scroll_offset: u32) -> RgbaImage {
         let mut img: ImageBuffer<Rgba<u8>, Vec<u8>> = ImageBuffer::new(width, height);
         for y in 0..height {
@@ -369,7 +370,7 @@ mod tests {
             let mut img = make_frame(TW, h, scroll);
             for y in content_h..h {
                 for x in 0..TW {
-                    let n = ((x as u32 * y as u32 + noise_seed) % 6) as u8;
+                    let n = ((x * y + noise_seed) % 6) as u8;
                     img.put_pixel(x, y, Rgba([n, n, n, 255]));
                 }
             }
@@ -459,7 +460,7 @@ mod tests {
             let mut img = make_frame(TW, h, scroll);
             for y in content_h..h {
                 for x in 0..TW {
-                    let n = ((x as u32 * y as u32 + noise_seed) % 6) as u8;
+                    let n = ((x * y + noise_seed) % 6) as u8;
                     img.put_pixel(x, y, Rgba([n, n, n, 255]));
                 }
             }
@@ -512,7 +513,7 @@ mod tests {
             let mut img = make_frame(TW, h, scroll);
             for y in content_rows..h {
                 for x in 0..TW {
-                    let n = ((x as u32 * y as u32 + noise_seed) % 6) as u8;
+                    let n = ((x * y + noise_seed) % 6) as u8;
                     img.put_pixel(x, y, Rgba([n, n, n, 255]));
                 }
             }
